Allow to run proof for the offload cases ---
### Motivation

We want to run the streaming proof for the offloaded data.
For the offloaded data, we can using the catch up read mode
because the offloaded status is on the ledger. We need to
cumulate some data for preparing the offload. Then reading
them from beginning to verify.

So If you want to verify the offloading data, delay the consumer
for a while, then request the proof with verifyOffloading config.